pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standart"], version = "^0.27.1"}
sqlalchemy = "^2.0.26"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.6.1"}
bcrypt = "^4.1.2"
phonenumbers = "^8.13.30"
pydantic-extra-types = "^2.5.0"
sqlalchemy-easy-softdelete = "^0.8.3"
libgravatar = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
